fix(phase-359 W10 fallout): drop a cast the refactor made redundant
`b6d7cd097` moved the thread and spin-loop cluster to the platform, which
changed `timeout_ms` to `u64` and left `ms as u64` behind. `-D warnings` rejects
it as `unnecessary_cast`, so `check-test-targets` was red on main.

The lint class CLAUDE.md flags, arriving by a type change rather than a
toolchain bump — and visible only in the lane that compiles this crate.

+++ packages/core/nros-node/src/executor/spin.rs
@@ -6433,7 +6433,7 @@ impl<'s> Executor<'s> {
         // never; the honest reading is "no deadline", which is what an
         // untimed `spin_blocking` already is.
         let start_us = self.now_us();
-        let timeout_us = opts.timeout_ms.map(|ms| ms as u64 * 1_000);
+        let timeout_us = opts.timeout_ms.map(|ms| ms * 1_000);
         let mut total_callbacks = 0usize;
 
         self.halt_flag
